requirements.txt
requests~=2.30.0
python-dotenv~=1.0.0
beautifulsoup4~=4.12.2